Extract Charty::ColorMethods module
