+++ lib/charty.rb
@@ -2,6 +2,7 @@ require_relative "charty/version"
 
 require_relative "charty/backends"
 require_relative "charty/colors"
+require_relative "charty/color_methods"
 require_relative "charty/plotter"
 require_relative "charty/layout"
 require_relative "charty/linspace"
